refactor: split internal packages into backend, and tui
