+++ internal/backend/collections/manager.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"database/sql"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/database"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/database"
 	"github.com/maniac-en/req/internal/log"
 )
