+++ internal/backend/collections/models.go
@@ -3,8 +3,8 @@ package collections
 import (
 	"time"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/database"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/database"
 )
 
 type CollectionEntity struct {
