+++ internal/backend/endpoints/manager_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"testing"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/testutils"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/testutils"
 )
 
 func TestEndpointsManagerCRUD(t *testing.T) {
