+++ internal/backend/endpoints/models.go
@@ -3,8 +3,8 @@ package endpoints
 import (
 	"time"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/database"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/database"
 )
 
 type EndpointEntity struct {
