+++ internal/backend/history/manager.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/database"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/database"
 	"github.com/maniac-en/req/internal/log"
 )
