+++ internal/backend/history/models.go
@@ -4,8 +4,8 @@ package history
 import (
 	"time"
 
-	"github.com/maniac-en/req/internal/crud"
-	"github.com/maniac-en/req/internal/database"
+	"github.com/maniac-en/req/internal/backend/crud"
+	"github.com/maniac-en/req/internal/backend/database"
 )
 
 type HistoryManager struct {
