+++ internal/tui/tabs/components.go
