+++ internal/tui/tabs/tab.go
