+++ main.go
@@ -12,13 +12,13 @@ import (
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/maniac-en/req/global"
-	"github.com/maniac-en/req/internal/app"
-	"github.com/maniac-en/req/internal/collections"
-	"github.com/maniac-en/req/internal/database"
-	"github.com/maniac-en/req/internal/endpoints"
-	"github.com/maniac-en/req/internal/history"
-	"github.com/maniac-en/req/internal/http"
+	"github.com/maniac-en/req/internal/backend/collections"
+	"github.com/maniac-en/req/internal/backend/database"
+	"github.com/maniac-en/req/internal/backend/endpoints"
+	"github.com/maniac-en/req/internal/backend/history"
+	"github.com/maniac-en/req/internal/backend/http"
 	"github.com/maniac-en/req/internal/log"
+	"github.com/maniac-en/req/internal/tui/app"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/pressly/goose/v3"
 )
